+++ tests/ValueObject/DomainObjectTest.php
@@ -15,16 +15,65 @@ it('can create a domain object from resolvable class names', function (string $c
         ->namespace->toEqual($relativeNamespace)
         ->path->toEqual($expectedPath);
 })->with([
-    ['Domain\Invoicing\Models\Invoice', 'Invoicing', 'Models', 'Invoice'],
-    ['Domain\Invoicing\Models\Payment\InvoicePayment', 'Invoicing', 'Models\Payment', 'InvoicePayment'],
-    ['Domain\Internal\Invoicing\Models\Invoice', 'Internal\Invoicing', 'Models', 'Invoice'],
-    ['Domain\Internal\Invoicing\Models\Payment\InvoicePayment', 'Internal\Invoicing', 'Models\Payment', 'InvoicePayment'],
-    ['Domain\Invoicing\AdHoc\Thing', 'Invoicing', 'AdHoc', 'Thing'],
-    ['Domain\Invoicing\AdHoc\Nested\Thing', 'Invoicing', 'AdHoc\Nested', 'Thing'],
+    [
+        // Full Class Name
+        'Domain\Invoicing\Models\Invoice',
+
+        // Domain
+        'Invoicing',
+
+        // Object Namespace
+        'Models',
+
+        // Object Name
+        'Invoice',
+    ],
+
+    [
+        'Domain\Invoicing\Models\Payment\InvoicePayment',
+        'Invoicing',
+        'Models',
+        'Payment\InvoicePayment',
+    ],
+
+    [
+        'Domain\Internal\Invoicing\Models\Invoice',
+        'Internal\Invoicing',
+        'Models',
+        'Invoice',
+    ],
+
+    [
+        'Domain\Internal\Invoicing\Models\Payment\InvoicePayment',
+        'Internal\Invoicing',
+        'Models',
+        'Payment\InvoicePayment',
+    ],
+
+    [
+        'Domain\Invoicing\AdHoc\Thing',
+        'Invoicing',
+        '',
+        'AdHoc\Thing',
+    ],
+
+    [
+        'Domain\Invoicing\AdHoc\Nested\Thing',
+        'Invoicing',
+        '',
+        'AdHoc\Nested\Thing',
+    ],
+
+    [
+        'Domain\Invoicing\InvoicingServiceProvider',
+        'Invoicing',
+        '',
+        'InvoicingServiceProvider',
+    ],
 
     // Ad-hoc objects inside subdomains are not supported for now
-    // ['Domain\Internal\Invoicing\AdHoc\Thing', 'Internal\Invoicing', 'AdHoc', 'Thing'],
-    // ['Domain\Internal\Invoicing\AdHoc\Nested\Thing', 'Internal\Invoicing', 'AdHoc\Nested', 'Thing'],
+    // ['Domain\Internal\Invoicing\AdHoc\Thing', 'Internal\Invoicing', '', 'Adhoc\Thing'],
+    // ['Domain\Internal\Invoicing\Deeply\Nested\Adhoc\Thing', 'Internal\Invoicing', '', 'Deeply\Nested\Adhoc\Thing'],
 ]);
 
 it('cannot create a domain object from unresolvable classes', function (string $class) {